Removing decode-raw and implementing it in file_peeking
